src/requirements_dev.txt
# Base website
-r ./requirements.txt

# Tools for maintaining requirements.txt:
pip-tools == 6.1.0
pur == 5.4.1

# Useful extra developer packages:
black == 20.8b1
isort == 5.8.0
pytest-sugar == 0.9.4
termcolor >= 1.1.0  # for pytest-sugar
pre-commit == 2.13.0

# Debugging
django-debug-toolbar == 3.2.1
django-extensions == 3.1.2

# graphql
graphene-django == 3.0.0b7
